+++ src/grammar/ast.ts
@@ -214,8 +214,17 @@ export type ASTField = {
     id: number,
     name: string,
     type: ASTTypeRef,
-    init: bigint | boolean | string | null | undefined,
+    init: ASTExpression | null,
     as: string | null,
+    ref: ASTRef
+}
+
+export type ASTConstant = {
+    kind: 'def_constant',
+    id: number,
+    name: string,
+    type: ASTTypeRef,
+    value: ASTExpression,
     ref: ASTRef
 }
 
@@ -227,7 +236,7 @@ export type ASTContract = {
     name: string,
     traits: ASTString[],
     attributes: ASTContractAttribute[],
-    declarations: (ASTField | ASTFunction | ASTInitFunction | ASTReceive)[],
+    declarations: (ASTField | ASTFunction | ASTInitFunction | ASTReceive | ASTConstant)[],
     ref: ASTRef
 }
 
@@ -363,7 +372,7 @@ export type ASTStatementRepeat = {
 
 export type ASTStatement = ASTStatementLet | ASTStatementReturn | ASTStatementExpression | ASTSTatementAssign | ASTCondition | ASTStatementWhile | ASTStatementUntil | ASTStatementRepeat;
 export type ASTExpression = ASTOpBinary | ASTOpUnary | ASTOpField | ASTNumber | ASTID | ASTBoolean | ASTOpCall | ASTOpCallStatic | ASTOpNew | ASTNull | ASTLvalueRef | ASTInitOf | ASTString;
-export type ASTNode = ASTExpression | ASTProgram | ASTStruct | ASTField | ASTContract | ASTArgument | ASTFunction | ASTOpCall | ASTStatementLet | ASTStatementReturn | ASTProgram | ASTPrimitive | ASTOpCallStatic | ASTStatementExpression | ASTNativeFunction | ASTSTatementAssign | ASTOpNew | ASTNewParameter | ASTTypeRef | ASTNull | ASTCondition | ASTInitFunction | ASTStatementWhile | ASTStatementUntil | ASTStatementRepeat | ASTReceive | ASTLvalueRef | ASTString | ASTTrait | ASTProgramImport | ASTFunction | ASTNativeFunction | ASTInitOf | ASTString;
+export type ASTNode = ASTExpression | ASTProgram | ASTStruct | ASTField | ASTContract | ASTArgument | ASTFunction | ASTOpCall | ASTStatementLet | ASTStatementReturn | ASTProgram | ASTPrimitive | ASTOpCallStatic | ASTStatementExpression | ASTNativeFunction | ASTSTatementAssign | ASTOpNew | ASTNewParameter | ASTTypeRef | ASTNull | ASTCondition | ASTInitFunction | ASTStatementWhile | ASTStatementUntil | ASTStatementRepeat | ASTReceive | ASTLvalueRef | ASTString | ASTTrait | ASTProgramImport | ASTFunction | ASTNativeFunction | ASTInitOf | ASTString | ASTConstant;
 export type ASTType = ASTPrimitive | ASTStruct | ASTContract | ASTTrait;
 
 type DistributiveOmit<T, K extends keyof any> = T extends any
@@ -451,6 +460,14 @@ export function traverse(node: ASTNode, callback: (node: ASTNode) => void) {
             traverse(e, callback);
         }
     }
+    if (node.kind === 'def_field') {
+        if (node.init) {
+            traverse(node.init, callback);
+        }
+    }
+    if (node.kind === 'def_constant') {
+        traverse(node.value, callback);
+    }
 
     //
     // Statements
@@ -529,10 +546,8 @@ export function traverse(node: ASTNode, callback: (node: ASTNode) => void) {
         for (let e of node.args) {
             traverse(e, callback);
         }
-        return;
     }
     if (node.kind === 'new_parameter') {
         traverse(node.exp, callback);
-        return;
     }
 }